requirements.txt
django
djangorestframework
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
python-decouple
django-rest-passwordreset
celery
redis